move Header to the reusableComponents folder
